Renamed tests/utils/utils.py to tests/utils/channel.py
`raiden_contracts/tests/utils/utils.py` only contains utility functions for channels, especially related to the settlement function. It was renamed to `channel.py` for clarity.
