+++ raiden_contracts/tests/utils/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
 
-from .contracts import *
-from .utils import *
 from .config import *
+from .contracts import *
+from .channel import *
