+++ raiden_contracts/tests/utils/channel.py
